Create shared-icons module
