+++ settings.gradle.kts
@@ -32,3 +32,4 @@ dependencyResolutionManagement {
 
 include(":composeApp")
 include(":androidApp")
+include(":shared-icons")
